# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GramPulse API
  description: API for GramPulse Dashboard, including Copilot, Market, Intelligence, and Operations.
  version: 1.0.0
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /copilot/chat:
    post:
      summary: AI Copilot Chat Interaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                conversationId:
                  type: string
                content:
                  type: string
                contextType:
                  type: string
                contextId:
                  type: string
      responses:
        '200':
          description: AI Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      conversationId:
                        type: string
                      messageId:
                        type: string
                      content:
                        type: string
                      suggestedPrompts:
                        type: array
                        items:
                          type: string
                      sources:
                        type: array
                        items:
                          type: object
                      actions:
                        type: array
                        items:
                          type: object
                  meta:
                    type: object
                  error:
                    type: string
                    nullable: true